[*] move scheduler.hpp to impl/
