+++ include/co/net/tcp_stream.hpp
@@ -4,11 +4,11 @@
 #include <span>
 
 #include <co/event.hpp>
+#include <co/impl/scheduler.hpp>
 #include <co/impl/uv_handler.hpp>
 #include <co/net/address.hpp>
 #include <co/net/status_codes.hpp>
 #include <co/result.hpp>
-#include <co/scheduler.hpp>
 #include <uv.h>
 
 namespace co::net
